fix(maskedinput): fix placeholder on focus
IF-1495
