+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Idle Focus Blur With Prefix/idle, focus, blur/chrome/focused.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:21f3f1cf8525390efe305f4ff141cfa3b6c7caf9b2bb269c39da4563c50d5b00
+size 600
