+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Idle Focus Blur With Prefix/idle, focus, blur/firefox/blured.png
