+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Idle Focus Blur With Prefix/idle, focus, blur/firefox/focused.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:655c51dae7aefd6c7a3db43747af195303ba6e7f296f638341eb71f126bca1dd
+size 727
