+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Idle Focus Blur/idle, focus, blur/chrome/focused.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:efdb7dd9de47f6142c1b9e84c69f2d288ff57cd102875a70f01c122bfa4a732d
-size 689
+oid sha256:8abf46eb841a89340427f2b8bcf380a3da2ea02a7a878293bcd830c9bae6468b
+size 388
