+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Idle Focus Blur/idle, focus, blur/chrome/idle.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:c3485d0ddd5bf718733f3763be8a48893ad45717471313a2476249b569764f07
-size 673
+oid sha256:1a0c90a3bdd2f4164472920d330c04e440f01eedaf173aefeb4b27f2e93bee22
+size 982
