+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Idle Focus Blur/idle, focus, blur/firefox/focused.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:88000fed599b431b8e964acb42b46ca9ba033e3753fb3f25b313e19d71ddd2b7
-size 792
+oid sha256:03285ad4de563846a6716f06837098eea3520de03445a938487b2844807609a1
+size 547
