+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Idle Focus Blur/idle, focus, blur/firefox/idle.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:7b2a8382f32ef0315319630f90e80684b45a44db51518758d671417df21d2fb1
-size 756
+oid sha256:2cc67b4aafcee5584161f1d0326c69f7176e4858517bb54bf47db2e0139767f1
+size 1054
