+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Select All By Prop/Plain focused/chrome/focused.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:26a5be67a19c7a78f345344b1658e858906e514c7b61a35bdd76ef7bb3408aeb
-size 1400
+oid sha256:b6bbfedadcc1701bff71d20950166aabb96b522ee3cb37486a85710702d00e1a
+size 1412
