+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Select All By Prop/Plain focused/firefox/focused.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:b2d7183abbb7bb20a3de43dca581985c9a76d0786f06d3492c5c0e601ad3731f
-size 1651
+oid sha256:6e076575b0f6a1a7541099764cc1d914ebdc56c25eb12878befd2e32095b6ada
+size 1812
