+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Select all by button/Plain, selected/chrome/selected.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:451dd4c9072f5baef55e3a2ba71fd9a49f678b6bc7fdd2114f2346f4fbfb308f
-size 2224
+oid sha256:3dab0c262a11b3e68bf84bd900ea30c4827b9899d8179f308f2cb405598a5451
+size 2220
