+++ packages/react-ui/.creevey/images/MaskedInput/Functional tests/Select all by button/Plain, selected/firefox/selected.png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:587c6200464a9f8e010b0ce7fdd9b6a3ef2bcadf421edcce47bb760a971967ea
-size 2589
+oid sha256:09ca04d27512cb7b7bd4dedef60b9afd039ff0292b4b0499f8ae2bf3309f4b58
+size 2665
